Move third-party invites to a separate package and doc' it
